Move PageSection to separate file
Docusaurus is weird and forces you to use `process.cwd()` because of
some implementation details...

<https://docusaurus.io/docs/en/api-pages#page-require-paths>
