+++ website/pages/en/index.js
@@ -33,23 +33,7 @@ man("")  # error: Method \`man\` does not exist
 \`\`\`
 `;
 
-// TODO(jez) No inline styles
-const PageSection = (props) => (
-  <div
-    style={{
-      backgroundColor: props.gray
-        ? '#f5f5f5'
-        : props.lightPurple
-        ? '#6255b0'
-        : 'white',
-      color: props.lightPurple ? 'white' : 'inherit',
-      padding: props.short ? '2em 0' : '3em 0',
-    }}
-    className="pageSection"
-  >
-    <div className="wrapper">{props.children}</div>
-  </div>
-);
+const PageSection = require(`${process.cwd()}/core/PageSection.js`);
 
 // TODO(jez) No inline style
 const Button = (props) => (
